+++ cogs/general.py
@@ -4,6 +4,7 @@ import aiohttp
 import inspect
 import os
 from discord.ext import commands
+from typing import Union
 
 
 class General(commands.Cog, name="General"):
@@ -92,6 +93,15 @@ class General(commands.Cog, name="General"):
                 f"**{country.capitalize()}** *({r['countryInfo']['iso3']})*",
                 embed=embed
             )
-    
+
+    @commands.command()
+    async def avatar(self, ctx, member: discord.User):
+        """Sends the member's discord avatar
+        """
+        embed = discord.Embed(title=f"{member.name}'s avatar:", color=0x8368ff)
+        embed.set_image(url=member.avatar_url)
+        await ctx.send(embed=embed)
+
+        
 def setup(bot):
     bot.add_cog(General(bot))
